Create and operator that can create and delete Couchbase clusters
This is the first commit of the Couchbase Operator. This commit
adds functionality to register a CRD with Kubernetes and deploy
the Couchbase Operator. The Couchbase Operator is able to then
deploy a cluster (pods and service) and then delete that cluster.

One limitation right now is that only one node will be provisioned
when creating the cluster since the functionality to cluster
multiple nodes has not yet been added.