Adds job package for cluster configuration
Jobs are async operations started by the operator with the
goal of conforming the couchbase cluster into its desired
state.  The job itself is composed of a container running the
couchbase-cli.  Since Jobs are native kubernetes constructs
they can also be ran standalone which allows for admins to do
post-init customization (see examples/node-init-job.yaml).

Jobs are ran via scheduler which provides a watchable channel
to the operator for notifications of jobs.  For instance, the
operator schedules a node-init job and when it's notified for
completion, the operator can adjust it's spec status accordingly.

Requires building cli-container (not currently in dockerhub):
  docker build --build-arg BRANCH=watson -f Dockerfile.couchbase-cli -t couchbase/couchbase-cli .