Ensure that the operator picks up where it left off when restarted
When the operator is restarted it will get an add update for each
cluster it is managing. Since the operator is restarted the clusters
it receives updates for likely already exist. In this case we need
to ensure that the cluster creation step is a no-op. We can tell if
the cluster has already been created if the cluster status is marked
as "runnning".

Change-Id: I385314d3b5d85ff995844456f36fe07b73e40c58
Reviewed-on: http://review.couchbase.org/83747
Tested-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>