Automatically provision nodes based on the size of the cluster
This commit ensures that the right number of pods are created
or removed based on changes in the desired state of the cluster.

We also track the members in the cluster in the config's status
section so that we can ensure that we locate the correct cluster
pods when checking the cluster state.

Change-Id: I4c652346be13fc0574230337397928f744a63fd7
Reviewed-on: http://review.couchbase.org/83894
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>