memberCounter shouldn't be compared to Spec.Size
When cluster scales down the memberCounter remains
unchanged.  When a node is added back to the cluster
the memberCounter will then be higher than Spec.Size,
and since rebalance waited for the values to equal the
rebalance never happened.
This fix instead uses len(members) which is actual
added to cluster instead of memberCounter.

Change-Id: I4e462963dac4ea83dd1051d24d0f30b1b1623d9b
Reviewed-on: http://review.couchbase.org/84086
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Tommie McAfee <tommie@couchbase.com>