member.First should seek to highest counter
As the memberset becomes fragemented from adding
and removing nodes it's possible to have 3 members
all started with counters higher than the range of the
membersets' length...ie [0-2], when the first member is
really at 4.  By seeking to max possible counter then
First member of the set is guaranteed to be discovered.

Change-Id: I0e789a5faf143ac07a9d4e11a2ddcae7fdaa9be2
Reviewed-on: http://review.couchbase.org/84087
Tested-by: Tommie McAfee <tommie@couchbase.com>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>