Remove unknown nodes as part of the top level reconcile function
We should remove unknown nodes as the first thing we do because we
want to avoid connecting to them when we start reconciling the
cluster. For example, the first thing we want to do is check for
failures, but we don't want to have to deal with going to a node
that is not part of the cluster.

Change-Id: I4949a9a5d1c5473f4a1a2e4c11e88b3acc1d107d
Reviewed-on: http://review.couchbase.org/84446
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>