Add comment describing the peer service
As it's not very clear exactly how this service behaves
by looking at the code.  It's actually quite nuanced in that
cluster IP needs to be "None" to allow p2p resolution.
Previously the value was an empty string which caused
kubernetes to fill in ClusterIP of the spec with a IP address
which behaved a load balancer to a single IP preventing direct pod dns.

Change-Id: I5f7b503985e5193cbd387ea6533a650b01111120
Reviewed-on: http://review.couchbase.org/84525
Tested-by: Tommie McAfee <tommie@couchbase.com>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>