Upgrade to kubernetes 1.8
This change upgrades our codebase to Kubernetes 1.8 and integrates
some new code generation commands that come in the latest kubernetes
updates. The code generation allows us to more easily setup our
project for development of a CRD.

To generate the code just run 'make build'. This will pull down a
docker image with the code generation executables for Kubernetes
and will generate the code. Depending on what you change in the
project you may may need to regenerate some of the code.

This change will still work with Kubernetes 1.7.

Change-Id: Id2ace31b8b7468443947fc3acaae2ddba305ec4f
Reviewed-on: http://review.couchbase.org/83911
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>