Additional pod labels specifying couchbase services
Labeling pods with services being run will allow for
kubernetes services to expose a cross-sectional view
of the cluster to applications.  For instance, a
load balancer can be created among the data pods for
kv workloads or query nodes can be isolated for its
respective use case.

Labels are of the form:
  'couchbase_service_<type> : enabled'

example of using the selectors:

List all pods with data service enbled:
  kubectl get po -l 'couchbase_service_data'

Pods with data and fts  service in example cluster:
  kubectl get po -l 'couchbase_service_data,couchbase_service_fts,couchbase_cluster=cb-example'

Change-Id: I24ffdb13e0aec50298ace4cac3759e65d8250d58
Reviewed-on: http://review.couchbase.org/84654
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Tommie McAfee <tommie@couchbase.com>