Test operator reverts manual bucket-edit
Disables flush via rest api and expects the
operator to re-enable. The rest connection
occurs over NodePort service created by the
test.

Change also adds an api_util for further interaction
with couchbase along with a service_util for getting
client urls for use with the api.

(one minor side effect to operator is the conversion
of bucket type is made public)

Change-Id: Ic378acd41baaa1047ce421b8ee72e9be2c399a40
Reviewed-on: http://review.couchbase.org/85599
Reviewed-by: Korrigan Clark <korrigan.clark@couchbase.com>
Tested-by: Tommie McAfee <tommie@couchbase.com>