K8S-21: Improve reconciling of nodes in the cluster
This change provides better support for catching the various
failure cases and also corrects issues/scales the cluster faster
by adding/removing nodes all in the same iteration of the reconcile
loop.

We also take advantage of the updated conditions to provide a better
status message on the state of the cluster theough the describe
output.

Change-Id: Ia5b25445dcc029f0706aee8d120e197ba02bac79
Reviewed-on: http://review.couchbase.org/84788
Tested-by: Mike Wiederhold <mike@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>