Record events when the operator takes various actions
This allows us to log events in etcd. These events can then be seen
in the `describe` output of the couchbase cluster. These events will
help the user to understand what type of things have happened in the
clusters past. This change only adds events to succesful operations.
A seperate change should be added for failure conditions.

Change-Id: Ia4dab49fc15931bafdffada21accd62f101e7251
Reviewed-on: http://review.couchbase.org/85864
Tested-by: Mike Wiederhold <mike@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>