K8S-11: Basic CRD validation
This adds some basic validation for checking configurations when
loading them into Kubernetes. In particular this does three types
of check.

1) Ensures that the type is correct (eg. string, int)
2) Adds minimum values for some integer fields (eg. Memory quotas)
3) Checks for required fields

CRD validation is alpha in 1.8 and will be beta in 1.9. To enable
this feature in 1.8 you must enable the feature gate for CRD
validation when starting Kubernetes. You must then enable CRD
validation in the operator by specifying the --enable-validation
flag when starting the operator

CRD validation will also allow for setting default values in the
future. As of right now this feature will be added as an alpha
in Kubernetes 1.9 and will likely be beta or ga in the following
release.

Change-Id: I5d0fce4b0738a95fd7440bf6927d4b9bd404e30c
Reviewed-on: http://review.couchbase.org/86027
Tested-by: Mike Wiederhold <mike@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>