Add a DeadMember state to the node reconcile state machine
Handling dead members used to be in the Finished state, but this is
problematic because finished is the state we transition to in the
case of errors. This means we might skip some checks we need to do
before we remove dead members.

One example is when we have a down node we transition the state to
finished because we are waiting for auto-failover. If we handle the
dead members in the Finished state then we will remove this node
from the cluster too early and won't detect once the node has been
failed over because the failed over node will be placed into the
UnknownNodes list.

Change-Id: Ia69412cb057e77c6befb2f9f0c4e13d245ff65ba
Reviewed-on: http://review.couchbase.org/86114
Tested-by: Mike Wiederhold <mike@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>