K8S-42: Fail pod create if unschedulable
Previously, the operator would only check if pod start is ok.
This was problematic for later attempts to add pod to cluster.
This change fails pod create when it cannot be scheduled so that
reconcile loop can stop attempts to add.
The failed pod is deleted so that user can change memory quota and try again.

Change-Id: Iaa6592e276eb40b449a2ac824b7e1456eb58d858
Reviewed-on: http://review.couchbase.org/86405
Tested-by: Tommie McAfee <tommie@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>