Lower the bucket wait timeout in TestNegBucketAdd to 90 seconds
At the end of this test we want to make sure a bucket isn't created.
The old wait was 300 seconds which is really long since the operator
runs through the reconcile loop every 8 seconds. 90 seconds should
be long enough to make sure a bucket is not and will never be created
with the bad parameters specified.

Change-Id: Ic41fa5a7ea029b8376e8839a2fe7856e7c031e78
Reviewed-on: http://review.couchbase.org/87397
Tested-by: Mike Wiederhold <mike@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>