K8S-95: Create the operator as a deployment in test cases
Previously we were creating the deployment as a pod and that didn't
allow us to run test cases where we test the operator failing and
being restarted.

The operator is deployed by default using the example/deployment.yaml
file but this can be overridden with the DEPLOYMENTSPEC environment
variable when running make test and make test-indv on the command
line.

Change-Id: I02e806e2a026f0d23460cba3d0b33260fc661bb2
Reviewed-on: http://review.couchbase.org/87826
Tested-by: Mike Wiederhold <mike@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>
Reviewed-by: Korrigan Clark <korrigan.clark@couchbase.com>