Add Dependencies to the Makefile
One issue I had: first it's not obvious you have to run the glide
command from the contents of the TLD, you need to dig around the docs.
This patch automatically installs dependencies on a first build
(hopefully the shell command is generic enough to work (bash and zsh
should be fine).

Change-Id: I7d423bccbc51e799b7b9e17c01b242e1d2e46239
Reviewed-on: http://review.couchbase.org/87971
Reviewed-by: Mike Wiederhold <mike@couchbase.com>
Tested-by: Simon Murray <simon.murray@couchbase.com>