K8S-24: Online Upgrades
Few fixes and refactors first.  Events moved out of the state machine
and into the create/delete/rebalance functions so anything calling them
gets the benefit.  Made cluster initialization code use the generic
reconcile node creation code to deduplicate effort. Removed cluster state
logging code from the reconcile FSM as it felt out of place, moved to
util so it can be used for debug in other places.

Main upgrade code.  The cluster status has been updated to contain state
used by the process.  A new utility class parses version strings and
allows decomposition into an optional prefix and semver fields.  The
main routine controls whether reconcilliation is allowed, in future
these may be rationalized into a single FSM... An upgrade is performed
only on matching releases, if it is an upgrade and up to the next major
version.  Other than that if we are upgrading create a new node based on
the server spec of the existing victim, perform a swap rebalance and
delete the old node.

Change-Id: I926b24436aff68726c6d366d81dfe8e045b6ef41
Reviewed-on: http://review.couchbase.org/88366
Tested-by: Simon Murray <simon.murray@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>