K8S-178: Delete Garbage Collection
The only way we can exit the main cluster runloop is via an event that a
cluster has been deleted.  This leads to two issues, the first a
misleading message that we are deleting the failed cluster, and second
the call to update the custom resource status fails as Kubernetes has
already deleted it.

Garbage collection happens automatically due to Kubernetes seeing the
pods and services are managed by the Couchbase cluster which has been
deleted.

Change-Id: If91771480d146b04bcec82340d6ca0cc82589ed4
Reviewed-on: http://review.couchbase.org/88988
Tested-by: Simon Murray <simon.murray@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>