K8S-108: TLS Support
The major feature is that couchbaseutil has now changed into a struct
which contains username, password and TLS configuration, which means
fewer paramaters being passed around.  The client is cached in the
cluster structure, this allows us to catch cluster deletion events and
notify the client which has a blocking channel which is consumed by
retryutil.  On deletion the channel is closed, becomes non-blocking and
we can terminate the retry loop with a suitable error condition.  The
netutil functions have been updated to accept a context too which may
have an associated timeout.

As regards TLS itself when we spin up a Couchbase server pod the sever
wildcard secret is mounted via a volume, the CA cert is uploaded and the
certificate activated via the API.  A flag in the member is toggled so
that all further references to it are via https://:18091.  One caveat is
that node addition over TLS doesn't work, another is that sadly mutual
verification is not until 5.5.0, and even then requires certificate
authorization, we'd ideally like basic HTTP auth over TLS...

Change-Id: I51e1a48448618151807e7fc7a98d5e0c9ecc3e5d
Reviewed-on: http://review.couchbase.org/88771
Tested-by: Simon Murray <simon.murray@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>