K8S-130: Rebalance Error Checking
Addresses the case where a rebalance is stopped within the reconcile
state machine, it then kills ejected nodes which results in data loss
and cluster failure.

Basically moves the same functionality from updates down into the
common reconcile code so it is now shared with the main reconcile loop.
Tested by stopping rebalance in the UI after renaming a server group and
during an upgrade in both plaintext and TLS modes.

Change-Id: If470b98cb31aff11f5a77b4c335f6a1727a15f30
Reviewed-on: http://review.couchbase.org/90729
Reviewed-by: Mike Wiederhold <mike@couchbase.com>
Tested-by: Simon Murray <simon.murray@couchbase.com>