K8S-188: Add Rebalance Completion Events
Creates an event to notify that rebalance has successfully completed to
be used as a synchronization primative for testing and a general user
notification.  Additionally when we abort a rebalance for whatever
reason an new event is raised for this also which will enable testing of
K8S-130 for both reconcile and upgrade.

Change-Id: If101ff0e4f3be6c0a1da8104c4b116e8ed921ade
Reviewed-on: http://review.couchbase.org/91008
Tested-by: Simon Murray <simon.murray@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>