K8S-128: Ability to display warnings when running cbopctl
Some changes to the configuration may result in downtime for all or
part of the cluster. For example, changing the eviction policy or
the io priority in a bucket will cause the bucket to be restarted.
This will lead the bucket needing to be warmed up and that process
could take a significant amount of downtime. This change allows us
to print warnings to the user about these types of events.

Change-Id: If292cda3b9af49900e1674c033e18855353baf5e
Reviewed-on: http://review.couchbase.org/91199
Tested-by: Mike Wiederhold <mike@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>