K8S-217: Add a production build target
We need to br able to ship cbopctl for various operating systems. As
a result we need to build cbopctl for linux, os x, and windows. The
easiest way to do this is to cross-compile the binaries. Since we
only want to do this when creating official builds I've added a
"prod" target.

Change-Id: I90359d2834c20b59da718e1e5065d0166e877809
Reviewed-on: http://review.couchbase.org/91211
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Simon Murray <simon.murray@couchbase.com>