K8S-205: Manual Failover Non-Determinism
A couple issues are addressed here.  First when a node is manually
failed over its state is status=active, membership=inactiveFailed which
makes no sense, however it isn't captured by our cluster status, so
although flagged as managed it doesn't exist in any other subset making
it uncontrollable.  Add in the new state and error on new, unexpected
states.

The second issue is a race where Couchbase server either reports needing
a rebalance, most of the time the failed node is inactiveFailed, however
1/20 times it is still marked as active.  In the former case the node is
rebalanced out and deleted, in the latter it is rebalanced out, and
second time around is deleted as it is unclustered.  This requires a
node removed event to be added to the unclustered case in the reconcile
loop.

As cluster status updates are now synchronous the wait-until-healthy code
breaks, e.g. the cluster enters scale-up, a new node added, this
triggers the health check, which passes, before the rebalance begins and
the balanced condition is unset which would have look unhealty.  We now
define a table of all healthy conditions, one of which is that that we
cannot be scaling.

Finally I spotted a while ago that there are lots of duplications of
code which performs managed node deletion.  These have all now been
rationalized into the canonical method.

Change-Id: I0b81c557d0f990ce06dbf076403551499b19567a
Reviewed-on: http://review.couchbase.org/91247
Tested-by: Simon Murray <simon.murray@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>