K8S-216: Fix TCP Waits From Tests
WaitForPod was updated in the operator to wait until the Couchbase admin
port comes up to prevent spam from retry loops in the logs.  This won't
work external to the cluster (e.g. tests) as the host is unresolvable
and cause tests to fail.  Make these checks optional.

The test itself was racy so I've fixed that up to pass.

Change-Id: I003ff70b7d52844a69297c4eb0ac61e8e85f8ee5
Reviewed-on: http://review.couchbase.org/91252
Tested-by: Simon Murray <simon.murray@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>