K8S-226: Check Nodes Have Been Balanced In
Looking at some of the test assumptions I realized that I'd broken the
semantics of the node added event.  This led me to discover that if we
kill off a node pre-rebalance that goes unnoticed and we report being
balanced, when we aren't according to the specification.  This addresses
both of these problems.

Also, once and for all, fix the race conditions associated with raising
events.

Change-Id: I1d9521b1f7089a3de44f6705c1af9d5e4d54b617
Reviewed-on: http://review.couchbase.org/91520
Tested-by: Simon Murray <simon.murray@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>