K8S-227: Reformat Condition Status
At present the 'describe' output is a little difficult to grok as a list
of structs is rendered with the same indentation.  This separates the
condition type from the struct and instead uses it as a key into a
condition map, which results in a far easier association with data and a
condition.

Change-Id: Id7c0291036be46eb99e1ad04f09e52a0b8fc9dbc
Reviewed-on: http://review.couchbase.org/91610
Tested-by: Simon Murray <simon.murray@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>