K8S-260: Track and handle failed nodes that can be added back to the cluster
We forgot to add handling for the case where a node is failed over, but
can be added back to the cluster. One way to get into this situation is
if you just click the failover button on an already healthy node. For
now the operator can just remove these nodes from the cluster, but in
the future when we have persistent volumes this will be the case where
we handle node recovery.

Change-Id: If1fc6467108f3d9fa191e77be007a07681150116
Reviewed-on: http://review.couchbase.org/92288
Tested-by: Mike Wiederhold <mike@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>