K8S-275: Document Networking Requirements
A lot of beta testers are asking questions about Kubernetes networking
in general as well as specifically related to how Couchbase Server fits
in.  This document attempts to describe how to perform client
connections, either from within the same Kubernetes cluster as another
pod, or from without which is where a lot of the confusion comes from.
And rightly so!

Change-Id: Ie00a2ee261d646d41fcc9728fb2fd897b1a4c3f1
Reviewed-on: http://review.couchbase.org/92921
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Simon Murray <simon.murray@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>