K8S-266: Add ability to set a SecurityContext for all pods
We need to tell Kubernetes to run our container as the couchbase
user and this is done through the SecurityContext. The couchbase
is always UID 1000 so we can hardcode this value. In general users
should always want this SecurityContext configuration, but there
may be cases where the user wants to run their containers as root,
as a different user, or set some of SE Linux options.

Since this should be rarely used I think we should keep it undocumented,
but available in case someone needs it. It if isn't specified then
we will just default it to the appropriate value.

Change-Id: Ic1fbf9d006f95c4cd25374ea802dbb6459d5998a
Reviewed-on: http://review.couchbase.org/95201
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>