K8S-252: Fix External Bucket Edit Test
Upon examination this defect should expect bucket edit events, three to
be precise.  Previously it would perform edits then check the CRD was in
a sane state.  It was always in a sane state as it isn't synchronously
updated and thus the test did nothing.

Now it waits for each reconciliation, asserting a bucket edit event is
raised.  Unfortunately while waiting for the bucket event, the call to
Watch actually returns every event ever, which includes prior bucket
edit events, so it wasn't actually waiting.  This is fixed by rejecting
past events.

Finally when setting the IO priority it miraculously got revereted by a
prior reconcile, so I've added in some suitable delays to allow
Couchbase server to get itself into a stable state before performing the
next test.

Change-Id: I12500c6168dd41fe5ec0344b059108d9fe5cd3bb
Reviewed-on: http://review.couchbase.org/93283
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>