K8S-113: Failure During Scale Down
Test fails due to no events being defined, this adds them in.

Better handling of rebalance failures effectively has fixed the issue
with data loss.  One little problem is the node removal code looks at
the running pods, so tries to scale down even though we have a node
scheduled to be removed as it has failed.  This caused a double removal
and an addition in two sepparate reconcile cycles.  Removing the failed
node from the running set solves the issue.

Change-Id: If4c974a37c3dba69f028c053183a6e8989daad26
Reviewed-on: http://review.couchbase.org/93340
Reviewed-by: Mike Wiederhold <mike@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>