K8S-292: Fix Kill Operator Race
When waiting for the operator to come back up we currently wait 10s
before polling again.  The test waits for the operator to come up and
then for a bucket change to be reverted.  If the operator comes up and
reverts the bucket before we poll for operator readiness we will miss
the bucket edit event entirely and hang.

This reduces the poll delay to a safer 1s.

Change-Id: I9007f07e4e7ee76de756a053d68d8d2eacef402e
Reviewed-on: http://review.couchbase.org/93424
Reviewed-by: Mike Wiederhold <mike@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>