K8S-222: Down & Failed Events
At present when a pod becomes unavailable either through being killed or
Couchbase server stopping reponding there is zero context expressed by
the event interface.  We just see things being added, rebalanced and
removed e.g. not a lot for support to diagnose at a glance.

This adds in support for node-down and failed events.  As it may take a
while for auto-failover to occur we risk adding multiple events for down
nodes.  For this purpose we use the built-in event recorder type which
caches events and aggregates similar ones together.

It would be nice to use this for all events but alas it will break tests
by aggregating rebalance events together.

Change-Id: I69d4975c237483d8237999bdf525d5bedc91f3ad
Reviewed-on: http://review.couchbase.org/93935
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>