K8S-229: Remove Foreign Node Fix
First fix is to remove the pod creation from e2eutil as this is done
implicitly by k8sutil.  Next up when deleting the illegal node and
adding in a new one the rebalance failed due to the node not having been
failed over.  I figure it's best in this case just to do a swap
rebalance.  Finally rejigged the test code to use DNS rather than IP
addresses for testability.

Change-Id: I5ee203e63848092dad2ccc92693c8625005d4160
Reviewed-on: http://review.couchbase.org/93883
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>