K8S-259: Delta recovery of failed over nodes
If a node is failed over then a deleta recovery is performed
if the node is healthy so that it can be added back to the cluster.

In the case that the failed over node is unhealthy then it is
recreated prior to the recovery (if it has persistent volumes),
then rebalanced in via delta recovery when it comes back online.

Change-Id: I8a566774dae2d38d8cc564894191ecfed0b2ccc9
Reviewed-on: http://review.couchbase.org/93836
Reviewed-by: Mike Wiederhold <mike@couchbase.com>
Tested-by: Mike Wiederhold <mike@couchbase.com>