K8S-273: Recover from cluster or operator crash
If cluster crashes and all of its pods are down then recovery is started.
In the case that the operator has also crashed then the memberset is recreated
from any persistent volumes that are available.  Next a single recoverable
pod is started so that the cluster can continue reconciling and recovering
remaining pods.

Change-Id: I9259ff9abe6470d73818a523c260673cbc6213f8
Reviewed-on: http://review.couchbase.org/94492
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Tommie McAfee <tommie@couchbase.com>