K8S-347: Scheduled Pod Deletion
The scheduler stripes pods for a particular server class across the
defined server groups in order to maintain balance.  Under the hood the
algorithm actually just adds a new server to the least populous server
group.

When we delete pods we should also schedule deletion in order to
maintain the same balanced characteristic.  Given the server class we
wish to scale down we simply pick a server from the most populous server
group.  Much like the former algorithm we sort candidate server groups
and servers in order to acheive deterministic characteristics.

Change-Id: Ib521a874eaa45be4b846fbb79d968c0d931da388
Reviewed-on: http://review.couchbase.org/94906
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>