K8S-261: Per Cluster Logging and Inhibition
The original idea was that when the cluster deletion event comes in we
black hole all logging output as it's irrelevant and potentially
confusing for support.  However there were some logs apparently coming
back from the dead.  Seems that the retry code and couchbase client were
using the global logger which I've fixed by propagating the cluster
specific logger via the cluster context.

Change-Id: Ia84708850f4c7e195d1d08fc706818443dad6d75
Reviewed-on: http://review.couchbase.org/94963
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>