K8S-357: Anti Affinity Scaling Test Fix
Test needs to be aware of the member creation failed event, it can also
be optimized to look for this event rather than timing out.  Also of
note is that when the config is reverted the reconcile loop thinks
noting needs doing and we never clean up the scaling status.

I also finally scratched that itch to do with the reconcile state
machine.  It is now driven by a lookup table, each function returns an
error so we don't need a magical flag, and can propagate these up to the
top level alerter.  We may in future want to implement special classes
of error to alter the log level.  We also stop the main loop from
reloading pod state on a reconcile error as it will delete members we
are tracking (i.e. are managed) so their states don't magically become
unmanaged rather than failed over.

Change-Id: Ib854f21d59c4b17c8cd6a76aaa51beb1ef1e5e45
Reviewed-on: http://review.couchbase.org/95128
Reviewed-by: Mike Wiederhold <mike@couchbase.com>
Tested-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>