K8S-332: Support Tool
From a high level we have sets of resource types which we need to
collect and write to some backend.  Resources can be scoped to the
cluster e.g. nodes, or scoped to a namespace.  By default with scoped
resources we also constrain to the clusters we care about e.g. only pods
with the couchbase label will be collected.  We can constrain further by
explicitly stating the cluster(s).  We can loosen the constraints with
the all flag which will collect everything in sight.

Some resources may implicitly have other things associated with them
e.g. events and logs which are also collected.

Change-Id: Ib4f886cdb2cdbf1ed8859e49bbf7c6ac6e0e8263
Reviewed-on: http://review.couchbase.org/95818
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>