K8S-369: Upgrade CRD Validation
Two changes here, first the CRD schema is well out of date. Second, we
have a schema, use it to validate cluster types, limits and
enumerations. This has the unfortunate side-effect that JSON validation
doesn't support the notion of null pointers, meaning that the bucket
configuration has altered slightly implementation wise, the API is
unchanged.  The schema doesn't support nullable lists or maps, so we
need to set these to omitempty to prevent this being sent on the wire.

Change-Id: Ie1ff12897c0e5d5f75fac13f80087cf1165ec5cc
Reviewed-on: http://review.couchbase.org/95315
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>