K8S-366: Remove Foreign Node Fix
There was no code path to remove unmanaged pods so we now track and
delete the pods.  The test itself did a blocking resize, then waited for
a rebalance to start.  Obviously once the resize had completed we'd
already done the rebalance, so trigger the resize asynchronously, wait
for the rebalance (unhealthy condition), before waiting for healthy
state.

Change-Id: Ic275a7c74bdcaa75649cf5d265d92944b0138fbe
Reviewed-on: http://review.couchbase.org/95357
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Korrigan Clark <korrigan.clark@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>