K8S-390: Reconcile Server Groups
The UI allows modification of server groups outside of automation so we
need a method to undo anything which breaks the model.  I've moved the
server group code from node initialisation into the reconcile loop.
Conceptually the same as before, create server groups, prepare an update
based on existing servers but choosing the server group from the pod
label.  If the current location and required location differ we commit
the update.

Change-Id: I210b1b410f535825ea8d3ccea2ca015458969b67
Reviewed-on: http://review.couchbase.org/95674
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>