K8S-398: Validation Updates
Adds support for checking anti-affinity is immutable.  Running with it
off means pods may get scheduled on the same node, turning it on may
have disasterous consequences of pods being destroyed to fix the
constraints.  It doesn't thankfully but still pods will not be placed as
the specification orders.

Adds support to check that bucket names are unique as we use this as the
primary key when looking up buckets while reconciling.

Change-Id: I825fa80e5fe6ac531b4bd9892c8cdf7cbe3c2993
Reviewed-on: http://review.couchbase.org/95883
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>