K8S-400: Ignore Faulty Pods in Scheduler
And so the simple model becomes like a sudoku solver... When we detect a
pod has failed we will add in a new one and eject the old, then delete
the old pod.  Problem is - in the context of scheduling - that we make
the decision based on all scheduled pods for a cluster, we have no
intelligence of what the operator intends to do.  To solve this simply
ignore pods whose state isn't healthy.  I suspect in future the operator
should make a plan of what it indends to do, then carries it out with
the help of the scheduler and can share this knowledge.

Change-Id: I2435ff783c5c9aa07715ee764bb8c33417e36027
Reviewed-on: http://review.couchbase.org/96050
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>