K8S-417: Server Group Pod Failure
When a pod was deleted (thanks QE!) the server group reconciler thought
it was a missing label and refused to do anything.  This ignores missing
pods to allow further reconciliation.

Also the call site of the reconcile was wrong, I've moved to to after
pod creation and before rebalance to optimise the process.  Also if we
perform an update we also force a rebalance, possibly fixing K8S-411.

Change-Id: Ib20e8e406d962459c418a9fa71130ba49313bd2b
Reviewed-on: http://review.couchbase.org/96254
Reviewed-by: Tommie McAfee <tommie@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>