K8S-416: Fix Pod Numbering Race
If we deleted the highest numbered pod while the operator was not
running we would have used its index to allocate the replacement on
operator restart.  This would fail to be added to the cluster as its
name matched an existing member.  This causes quite a lot of spam in the
retry loop.

Instead of implicitly guessing the correct value to start at we annotate
the cluster resource with the current good value.  This still leaves a
minor issue whereby rebalances fail while the old pod fails over, but
fixes itself up eventually.

Another strategy would be to use Couchbase server as the source of
truth, so feel free to debate.

Change-Id: I354a6e61b90f31f197a8c212b2b81e0d5bf4b82a
Reviewed-on: http://review.couchbase.org/96262
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>