Fix storage path restoration after restart
When a persisted Pod is killed it is replaced with a new Pod
with the same volumes as the old Pod.  Ideally, all of the Pod's
configuration should be restored from the old config dir but
we've discovered that data paths are set in the couchdb dir under
'etc' which isn't persisted.  Also since we can't reset the data_paths
since the pod is already 'part of a cluster' then the path change
could result in loss of data.

The fix here is to persist the etc directory so that all storage paths
can be restored.  This involves starting an InitContainer prior to the
main container which is the exact same couchbase image so that we
can copy the 'etc' data over to the persisted volume.  This is so that
we don't lose 'etc' data when the same volume is mounted into the
Pods main container at runtime, which is necessary to persist etc.

Note: etc and the default path both use the same persisted volume via SubPaths
so no new volume is needed for this.

Change-Id: I311beca4913910244fda13dc35ae05bb0e7032bf
Reviewed-on: http://review.couchbase.org/96303
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>